refactor: combine all triggers into a single binary to speed up compilation
