+++ plugins/buildpacks/.gitignore
@@ -5,3 +5,4 @@
 /post-delete
 /post-extract
 /report
+/triggers
