+++ plugins/buildpacks/Makefile
@@ -3,7 +3,7 @@ include ../../common.mk
 GO_ARGS ?= -a
 
 SUBCOMMANDS = subcommands/add subcommands/clear subcommands/list subcommands/remove subcommands/report subcommands/set
-TRIGGERS = triggers/install triggers/post-delete triggers/post-extract triggers/report
+TRIGGERS = install post-delete post-extract report
 build-in-docker: clean
 	docker run --rm \
 		-v $$PWD/../..:$(GO_REPO_ROOT) \
@@ -12,7 +12,6 @@ build-in-docker: clean
 		bash -c "GO_ARGS='$(GO_ARGS)' make -j4 build" || exit $$?
 
 build: commands subcommands triggers
-	$(MAKE) triggers-copy
 
 commands: **/**/commands.go
 	go build $(GO_ARGS) -o commands src/commands/commands.go
@@ -28,10 +27,8 @@ clean:
 src-clean:
 	rm -rf .gitignore src triggers vendor Makefile *.go glide.*
 
-triggers: $(TRIGGERS)
+triggers: **/**/triggers.go $(TRIGGERS)
+	go build $(GO_ARGS) -o triggers src/triggers/triggers.go
 
-triggers/%: src/triggers/*/%.go
-	go build $(GO_ARGS) -o $@ $<
-
-triggers-copy:
-	cp triggers/* .
+$(TRIGGERS): 
+	ln -s triggers $@
